Minor UI tweak - display a "no results" message (instead of "Welcome to Timesketch") when a search of sketches has no results
The is a super minor UI tweak. Right now, you get the “Welcome to
Timesketch” screen if there are no results when searching for sketches,
which doesn’t make much sense. This pull request changes this behavior
to display the search bar, results table, and a “no results” message in
the results table if there are no results.