Surface celery task errors (#530)
* Mark timelines status with failure if ingestion error

* Refactor celery tasks

* Set timeline status

* Refactor celery tasks

* Refactor celery tasks

* Refactor celery tasks

* Update UI mapings

* Update UI mapings

* Mention the disksize plugin (#529)

Without it, you will end up:

There was an error loading a Vagrantfile. The file being loaded
and the error message are shown below. This is usually caused by
a syntax error.

Path: /Users/foo/timesketch/vagrant/Vagrantfile
Line number: 8
Message: RuntimeError: vagrant-disksize is not installed ("vagrant plugin install vagrant-disksize")

Not a big deal and it will tell you what to do anyway, but mentioning it lowers the bar to use it I hope.

* Return error

* Permit semicolons

* Mark timelines status with failure if ingestion error

* Refactor celery tasks

* Set timeline status

* Refactor celery tasks

* Refactor celery tasks

* Refactor celery tasks

* Update UI mapings

* Update UI mapings

* Return error

* Permit semicolons