Changed marks on lower chart to bars
Some values were being cut off as only the time range ony extended to
the start timestamp for a session (i.e. the last session on the chart
could not be viewed). Using a bar extending form start_timestamp to
end_timestamp solves this. x2Offset is used to make the bars more
visible.