Session chart (#974)
* Required for merging

* Required for merging

* Required for merging

* Vue component for the session chart

* Use nested session id attribute

* Changed marks on lower chart to bars

Some values were being cut off as only the time range ony extended to
the start timestamp for a session (i.e. the last session on the chart
could not be viewed). Using a bar extending form start_timestamp to
end_timestamp solves this. x2Offset is used to make the bars more
visible.

* Add message indicating if only some sessions were retrieved

* Rename roundedSessions -> processedSessions

* Change query string to use nested session id

* Fixed requested changes

Also changing the default time range to be the last year of the
timeline, not the last year from now.

* Adding TODO to make chart size responsive