Introduce .dockerignore
Prior to this change, the lack of a .dockerignore file would cause a
(lengthy) docker build whenever a git operation was performed regardless
of the files that had changed in the working tree.

This change adds .git directories to the .dockerignore file.  Also, the
docs and notebooks directories are added since they are not involved
with the build process.