Provide example commands docker mac shares (#1068)
The description might be okay for advanced users, but it might be handy if people could c&p stuff. The commands are from https://github.com/google/timesketch/issues/841 - so it seems at least one person had this issue before.

Let me know if that is something not wanted to be covered in the readme and / or more OS should be explained.