Properly split file extension
Error in code on line 126 where the file extension is grabbed. Currently splitting and grabbing [1] will only grab you the first item after the file name. This may not always be the extension. Plaso currently outputs the .plaso file as filename.something.something2.somethin3.plaso. To properly grab the file extension and prevent issues when uploading due to unknown file extension it should be grabbing the last item in the split [-1] which would always be the real extension.

IE:
let fileExtension = fileName.split('.')[-1]